Updated udon13/config.h to compile against latest qmk
I'm new to QMK, so this is something of a shot in the dark.

The version of this file included here does not compile against the latest QMK, as
set up per the instructions.  With a little examination of [the config.h](https://github.com/qmk/qmk_firmware/commits/master/keyboards/themadnoodle/noodlepad/config.h)
for the upstream noodepad, it looks like some of the directives used here have been
retired or moved elsewhere; I've copied that history over to the udon13's config.h,
using the values that were already present where it made sense.

I compiled this and installed it on my unit, and it appears to work the same as
the image it came with.